feat(api): move nonce_lab_cli
