+++ package/api/src/utils/mod.rs
@@ -2,4 +2,5 @@ pub mod email;
 pub mod error;
 pub mod hash;
 pub mod jwt;
+pub mod nonce_lab;
 pub mod time;
